VFS-3531 Deleted '.localhost' from hostnames (#36)
To address new chrome feature:
https://bugs.chromium.org/p/chromium/issues/detail?id=489973
that prevents working with url of format *.localhost.

We change the internal (in-container) host names in all scenarios by subtracting from them the '.localhost' suffix.
However, Onedata expects its services to be available under FQDN of minimum 3 parts (eg. `a.b.com`, the `b.com` will not work). A simple workaround presented here appends a dot `b.com.` to the address in onezone/oneprovider config. The resulting address is from the DNS point of view the same as `b.com` and Onedata services work corectly.