VFS-3790 Add config option that toggles cert auto-generation in onepanel, update docker images to rc6 (#41)
Options ONEPANEL_AUTOGENERATE_CERT and ONEPANEL_AUTOGENERATED_CERT_DOMAIN allow to toggle the functionality of onepanel to automatically generate web certificate on startup. The certificate is signed by OneDataTestWebServerCa, which is publicly available, so it by no means ensure safe connections. For that reason, use this option only for test purposes. The cert will be generated only if none is found under cert path.